src/requirements.txt
Django>4.2
django-widget-tweaks<1.4.12
django-ckeditor
pillow
readtime
six